new change

+++ controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../model/goalModel");
+const User = require("../model/userModel");
 
 // @ desc   Get goals
 // @route   Get  /api/goals
 // @access  private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({user:req.user.id});
   res.status(200).json(goals);
 });
 
@@ -21,6 +22,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id
   });
   res.status(200).json(goal);
 });
@@ -34,6 +36,20 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not Found");
+  }
+ 
+  // const user  = await User.findById(req.user.id)
+
+  // Check For user
+  if (!req.user) {
+     res.status(401)
+     throw new Error('User Not Found')
+  }
+
+  // Make sure only logged in user matches the Goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('User not Authorized')
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -53,7 +69,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(400);
         throw new Error("Goal not Found");
       }
+    
+      // const user  = await User.findById(req.user.id)
 
+  // Check For user
+  if (!req.user) {
+     res.status(401)
+     throw new Error('User Not Found')
+  }
+
+  // Make sure only logged in user matches the Goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('User not Authorized')
+  }
+       
        await goal.deleteOne()
 
   res.status(200).json({id: req.params.id});
